code refactor and package structure
